Authenticated Most of the REST API

+++ mysite/SocialApp/views.py
@@ -9,7 +9,7 @@ from rest_framework.views import APIView
 from .models import Author, Post, Comment, LikedPost, InboxItem
 from .admin import AuthorCreationForm
 
-from .utils import AuthorToJSON, PostToJSON, CommentToJSON, InboxItemToJSON
+from .utils import AuthorToJSON, PostToJSON, CommentToJSON, StringListToPostCategoryList, PostListToJSON, InboxItemToJSON
 
 from django.views import generic
 from django.urls import reverse_lazy
@@ -91,10 +91,13 @@ class AuthorEndpoint(APIView):
         """
         Handles GET requests
         """
+
+        # Check for bad URI
         author_id = kwargs.get('author_id', -1)
         if author_id == -1:
             return HttpResponse(status=400)
 
+        # Check that associated author exists
         try:
             author = Author.objects.get(pk=author_id)
         except Author.DoesNotExist:
@@ -102,6 +105,7 @@ class AuthorEndpoint(APIView):
         except Exception:
             return HttpResponse(status=400)
 
+        # Return public info about the author
         json = AuthorToJSON(author)
         if json:
             return JsonResponse(json)
@@ -112,11 +116,13 @@ class AuthorEndpoint(APIView):
         """
         Handles POST requests
         """
-        # TODO: Make this authenticated
+
+        # Check for bad URI
         author_id = kwargs.get("author_id", -1)
         if author_id == -1:
             return HttpResponse(status=400)
 
+        # Check that associated author exists
         try:
             author = Author.objects.get(pk=author_id)
         except Author.DoesNotExist:
@@ -124,6 +130,15 @@ class AuthorEndpoint(APIView):
         except Exception:
             return HttpResponse(status=400)
 
+        # Check that a user is authenticated
+        if not request.user.is_authenticated:
+            return HttpResponse(status=401)
+
+        # Check that the right user is authenticated
+        if request.user != author:
+            return HttpResponse(status=401)
+
+        # Update author info
         jsonData = request.data
         author.displayName = jsonData.get("displayName")
         author.github = jsonData.get("github")
@@ -140,7 +155,6 @@ class PostEndpoint(APIView):
         """
         Handles GET requests to retrieve a JSON representation of the post specified in the URL
         """
-        # TODO: Make this authenticated if not PUBLIC
         author_id = kwargs.get("author_id", -1)
         if author_id == -1:
             return HttpResponse(status=400)
@@ -165,6 +179,14 @@ class PostEndpoint(APIView):
 
         if post.author != author:
             return HttpResponse(status=404)
+
+        # Public posts can be viewed by anyone
+        if post.visibility != "PUBLIC":
+            # Check that a user is authenticated
+            if not request.user.is_authenticated:
+                return HttpResponse(status=401)
+
+            # TODO: Must be author's friend to have access
 
         json = PostToJSON(post)
         if json:
@@ -176,11 +198,13 @@ class PostEndpoint(APIView):
         """
         Handles POST requests to update the post specified in the URL
         """
-        # TODO: Make this authenticated
+        
+        # Check for bad URI (author portion)
         author_id = kwargs.get("author_id", -1)
         if author_id == -1:
             return HttpResponse(status=400)
 
+        # Make sure author exists
         try:
             author = Author.objects.get(pk=author_id)
         except Author.DoesNotExist:
@@ -188,10 +212,12 @@ class PostEndpoint(APIView):
         except Exception:
             return HttpResponse(status=400)
 
+        # Check for bad URI (post portion)
         post_id = kwargs.get("post_id", -1)
         if post_id == -1:
             return HttpResponse(status=400)
 
+        # Make sure post exists
         try:
             post = Post.objects.get(pk=post_id, author=author)
         except Post.DoesNotExist:
@@ -199,10 +225,18 @@ class PostEndpoint(APIView):
         except Exception:
             return HttpResponse(status=400)
 
+        # Make sure the author is actually the post's author
         if post.author != author:
             return HttpResponse(status=404)
 
-        # TODO: Handle categories
+        # Check that a user is authenticated
+        if not request.user.is_authenticated:
+            return HttpResponse(status=401)
+
+        # Check that the right user is authenticated
+        if request.user != author:
+            return HttpResponse(status=401)
+
         jsonData = request.data
         post.title = jsonData.get("title")
         post.description = jsonData.get("description")
@@ -210,6 +244,7 @@ class PostEndpoint(APIView):
         post.content = jsonData.get("content")
         post.visibility = jsonData.get("visibility")
         post.unlisted = bool(jsonData.get("unlisted"))
+        post.categories.set(StringListToPostCategoryList(jsonData.get("categories")))
         post.save()
 
         return HttpResponse(status=200)
@@ -218,11 +253,13 @@ class PostEndpoint(APIView):
         """
         Handles DELETE requests to delete the post specified in the URL
         """
-        # TODO: Make this authenticated
+
+        # Check for bad URI (author portion)
         author_id = kwargs.get("author_id", -1)
         if author_id == -1:
             return HttpResponse(status=400)
 
+        # Make sure author exists
         try:
             author = Author.objects.get(pk=author_id)
         except Author.DoesNotExist:
@@ -230,37 +267,46 @@ class PostEndpoint(APIView):
         except Exception:
             return HttpResponse(status=400)
 
+        # Check for bad URI (post portion)
         post_id = kwargs.get("post_id", -1)
         if post_id == -1:
             return HttpResponse(status=400)
 
+        # Make sure post exists
         try:
-            post = Post.objects.get(pk=post_id)
+            post = Post.objects.get(pk=post_id, author=author)
         except Post.DoesNotExist:
             return HttpResponse(status=404)
         except Exception:
             return HttpResponse(status=400)
 
+        # Make sure the author is actually the post's author
         if post.author != author:
             return HttpResponse(status=404)
+
+        # Check that a user is authenticated
+        if not request.user.is_authenticated:
+            return HttpResponse(status=401)
+
+        # Check that the right user is authenticated
+        if request.user != author:
+            return HttpResponse(status=401)
 
         post.delete()
 
         return HttpResponse(status=200)
 
-    # TODO: manage post creation based on content type
-    # TODO: Should put prevent overwriting an existing post???
     def put(self, request, *args, **kwargs):
         """
         Handles PUT requests to create a new post with the ID and author specified by the URL
         """
-        # TODO: Make this authenticated
-        # TODO: Manage how the post is created based on the content type
-        # TODO: Should this protect us from overwriting an existing post?
+        
+        # Check for bad URI (author portion)
         author_id = kwargs.get("author_id", -1)
         if author_id == -1:
             return HttpResponse(status=400)
 
+        # Make sure author exists
         try:
             author = Author.objects.get(pk=author_id)
         except Author.DoesNotExist:
@@ -268,14 +314,24 @@ class PostEndpoint(APIView):
         except Exception:
             return HttpResponse(status=400)
 
+        # Check for bad URI (post portion)
         post_id = kwargs.get("post_id", -1)
         if post_id == -1:
             return HttpResponse(status=400)
 
+        # Check for malformed URI (post portion)
         try:
             test = uuid.UUID(post_id)
         except:
             return HttpResponse(status=400)
+
+        # Check that a user is authenticated
+        if not request.user.is_authenticated:
+            return HttpResponse(status=401)
+
+        # Check that the right user is authenticated
+        if request.user != author:
+            return HttpResponse(status=401)
 
         try:
             jsonData = request.data
@@ -292,12 +348,12 @@ class AuthorPostsEndpoint(APIView):
     """
     The author/{AUTHOR_ID}/posts/ endpoint
     """
-    # TODO: Handle image posts
     def get(self, request, *args, **kwargs):
         """
         Handles GET requests to return the author's last N posts
         """
         # TODO: Paginate results and sort by date
+        # TODO: Authenticate? Or only allow this endpoint to return PUBLIC posts?
         author_id = kwargs.get("author_id", -1)
         if author_id == -1:
             return HttpResponse(status=400)
@@ -309,31 +365,35 @@ class AuthorPostsEndpoint(APIView):
         except Exception:
             return HttpResponse(status=400)
 
-        post_json_list = []
         posts = Post.objects.filter(author=author)
-        for post in posts:
-            json = PostToJSON(post)
-            if json:
-                post_json_list.append(json)
-        return JsonResponse({"posts":post_json_list})
+        json = PostListToJSON(posts)
+        return JsonResponse({"posts":json})
 
-    # TODO: manage post creation based on content type
     def post(self, request, *args, **kwargs):
         """
         Handles POST requests to create a new post for the author specified by the URL
         """
-        # TODO: Make this authenticated
-        # TODO: Manage how the post is created based on the content type
+        
+        # Check for bad URI
         author_id = kwargs.get("author_id", -1)
         if author_id == -1:
             return HttpResponse(status=400)
 
+        # Make sure author exists
         try:
             author = Author.objects.get(pk=author_id)
         except Author.DoesNotExist:
             return HttpResponse(status=404)
         except Exception:
             return HttpResponse(status=400)
+
+        # Check that a user is authenticated
+        if not request.user.is_authenticated:
+            return HttpResponse(status=401)
+
+        # Check that the right user is authenticated
+        if request.user != author:
+            return HttpResponse(status=401)
 
         try:
             jsonData = request.data
@@ -392,11 +452,13 @@ class PostCommentsEndpoint(APIView):
         """
         Handles POST requests to add a new comment to a post
         """
-        # TODO: Make this authenticated
+
+        # Check for bad URI (author portion)
         author_id = kwargs.get("author_id", -1)
         if author_id == -1:
             return HttpResponse(status=400)
 
+        # Make sure author exists
         try:
             author = Author.objects.get(pk=author_id)
         except Author.DoesNotExist:
@@ -404,10 +466,12 @@ class PostCommentsEndpoint(APIView):
         except Exception:
             return HttpResponse(status=400)
 
+        # Check for bad URI (post portion)
         post_id = kwargs.get("post_id", -1)
         if post_id == -1:
             return HttpResponse(status=400)
 
+        # Make sure post exists
         try:
             post = Post.objects.get(pk=post_id)
         except Post.DoesNotExist:
@@ -415,12 +479,17 @@ class PostCommentsEndpoint(APIView):
         except Exception:
             return HttpResponse(status=400)
 
+        # Make sure the author in the URI matches the post's author
         if post.author != author:
             return HttpResponse(status=404)
 
+        # Check that a user is authenticated so it can be the author of the post
+        if not request.user.is_authenticated:
+            return HttpResponse(status=401)
+
         try:
             jsonData = request.data
-            comment = Comment(author=author, post=post, comment=jsonData.get("comment"),
+            comment = Comment(author=request.user, post=post, comment=jsonData.get("comment"),
                               content_type=jsonData.get("contentType"))
             comment.save()
             return HttpResponse(status=200)
